Moved DTOs to dto package + more
