+++ src/main/java/com/busanit501/travelproject/repository/LocationJh1Repository.java
@@ -1,7 +1,7 @@
 package com.busanit501.travelproject.repository;
 
 import com.busanit501.travelproject.domain.Location;
-import com.busanit501.travelproject.service.manager.LocationValueJh1DTO;
+import com.busanit501.travelproject.dto.LocationValueJh1DTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
