+++ src/main/java/com/busanit501/travelproject/service/manager/ManagerJh1Service.java
@@ -1,5 +1,7 @@
 package com.busanit501.travelproject.service.manager;
 
+import com.busanit501.travelproject.dto.LocationValueJh1DTO;
+import com.busanit501.travelproject.dto.ProductJh1DTO;
 import jakarta.transaction.Transactional;
 
 import java.util.List;
